Renamed filehandler.py to files.py
